move dbmanager to other file
